fix:PR feedback on consts
