+++ src/constants/index.js
@@ -382,3 +382,12 @@ export const TIME_OF_DAY_PERIODS = [
     color: '#2ec27e'
   }
 ];
+
+export const TRACKING_CONTROL_STATES = {
+  FULLY_HEATMAPPED: 'FULLY_HEATMAPPED',
+  PARTIALLY_HEATMAPPED: 'PARTIALLY_HEATMAPPED',
+  FULLY_PINNED: 'FULLY_PINNED',
+  PARTIALLY_PINNED: 'PARTIALLY_PINNED',
+  FULLY_VISIBLE: 'FULLY_VISIBLE',
+  PARTIALLY_VISIBLE: 'PARTIALLY_VISIBLE',
+};
